Revert "refactor(JAQPOT-467): rename model type to OPENAI_LLM (#133)"
This reverts commit b155d31bec7f2f5a9839c49e80dbb76d8661824c.
