+++ src/main/kotlin/org/jaqpot/api/mapper/ModelTypeMapper.kt
@@ -23,7 +23,7 @@ fun ModelTypeDto.toEntity(): ModelType {
         ModelTypeDto.QSAR_TOOLBOX_QSAR_MODEL -> ModelType.QSAR_TOOLBOX_QSAR_MODEL
         ModelTypeDto.QSAR_TOOLBOX_PROFILER -> ModelType.QSAR_TOOLBOX_PROFILER
         ModelTypeDto.DOCKER -> ModelType.DOCKER
-        ModelTypeDto.OPENAI_LLM -> ModelType.OPENAI_LLM
+        ModelTypeDto.DOCKER_LLM -> ModelType.DOCKER_LLM
     }
 }
 
@@ -47,6 +47,6 @@ fun ModelType.toDto(): ModelTypeDto {
         ModelType.QSAR_TOOLBOX_QSAR_MODEL -> ModelTypeDto.QSAR_TOOLBOX_QSAR_MODEL
         ModelType.QSAR_TOOLBOX_PROFILER -> ModelTypeDto.QSAR_TOOLBOX_PROFILER
         ModelType.DOCKER -> ModelTypeDto.DOCKER
-        ModelType.OPENAI_LLM -> ModelTypeDto.OPENAI_LLM
+        ModelType.DOCKER_LLM -> ModelTypeDto.DOCKER_LLM
     }
 }
