+++ src/main/kotlin/org/jaqpot/api/service/prediction/runtime/runtimes/JaqpotPyV6Runtime.kt
@@ -1,10 +1,9 @@
-package org.jaqpot.api.service.prediction.runtime.runtimes.rest
+package org.jaqpot.api.service.prediction.runtime.runtimes
 
 import org.jaqpot.api.model.DatasetDto
 import org.jaqpot.api.model.PredictionModelDto
 import org.jaqpot.api.model.PredictionRequestDto
 import org.jaqpot.api.service.prediction.runtime.config.RuntimeConfiguration
-import org.jaqpot.api.service.prediction.runtime.runtimes.RuntimeBase
 import org.springframework.stereotype.Component
 
 @Component
