--- src/main/resources/db/migration/V47__rename_llm_model_type_to_openai.sql
@@ -1,3 +1,0 @@
-UPDATE model
-SET type = 'OPENAI_LLM'
-WHERE type = 'DOCKER_LLM';
